requirements.txt
absl-py==1.4.0
alembic==1.11.1
aniso8601==9.0.1
astunparse==1.6.3
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dlib==19.24.1
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-MySQL==1.5.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.9
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.10
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyMySQL==1.0.3
python-dateutil==2.8.2
python-version==0.0.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.15
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
# tensorflow-io-gcs-filesystem==0.32.0
# termcolor==2.3.0
# typing_extensions==4.6.1
# tzdata==2023.3
# urllib3==1.26.16
# Werkzeug==2.3.4
# wrapt==1.14.1
